feat: add search page

+++ src/app/api/trips-reservations/trip-cancel/[tripReservationId]/route.ts
@@ -2,10 +2,21 @@ import { deleteTripReservation } from "@/services/trips";
 import { NextResponse } from "next/server";
 
 export async function DELETE(
-  request: Request,
-  { params }: { params: { tripReservationId: string } }
+  _request: Request,
+  {
+    params,
+  }: {
+    params: { tripReservationId: string };
+  }
 ) {
-  const tripReservationId = params.tripReservationId;
+  const tripReservationId = params?.tripReservationId;
+
+  if (!tripReservationId) {
+    return NextResponse.json(
+      { message: "ID da reserva não existe!" },
+      { status: 404 }
+    );
+  }
 
   const trip = await deleteTripReservation({ tripReservationId });
